✨ feat: athena trigger
