+++ .gitignore
@@ -6,6 +6,8 @@ __pycache__/
 # C extensions
 *.so
 
+dag_test/*
+
 ## logs file
 logs/*
 data/*
