+++ requirements.txt
@@ -24,3 +24,4 @@ yarl==1.9.2
 zipp==3.15.0
 asyncio==3.4.3
 fake-useragent==1.1.3
+apache-airflow-providers-amazon
